Moved `ProfileTabs` to tabs package
